requirements.txt
asgiref==3.7.2
Django==5.0.1
pillow==10.2.0
pip==23.3.2
setuptools==69.0.3
sqlparse==0.4.4
typing_extensions==4.9.0
wheel==0.42.0
django-jalali-date==1.1.0